// src/main/enums/enums.ts
export enum Direction {
  LEFT_FROM_NORTH = 'W',
  RIGHT_FROM_NORTH = 'E',
  LEFT_FROM_EAST = 'N',
  RIGHT_FROM_EAST = 'S',
  LEFT_FROM_SOUTH = 'E',
  RIGHT_FROM_SOUTH = 'W',
  LEFT_FROM_WEST = 'S',
  RIGHT_FROM_WEST = 'N',
}

export enum Compass {
  NORTH = 'N',
  EAST = 'E',
  SOUTH = 'S',
  WEST = 'W',
}